Base.show — Method

Base.show(io::IO, pe::PositionEncoding)

Displays the PositionEncoding object.

Arguments

io::IO: The I/O stream.

pe::PositionEncoding: The PositionEncoding object.

VDJDL.Embeddings.make_position_encoding — Function

make_position_encoding(dim_embedding::Int, seq_length::Int, n::Int=10000)

Generates a positional encoding matrix.

Arguments

dim_embedding::Int: The dimension of the embedding.

seq_length::Int: The length of the sequence.

n::Int: A scaling factor. Defaults to 10000.

Returns

Matrix{Float32}: A matrix containing positional encodings.

VDJDL.Embeddings.PositionEncoding — Type

PositionEncoding(dim_embedding::Int, max_length::Int=1000)

Constructs a PositionEncoding object with specified embedding dimension and maximum sequence length.

Arguments

dim_embedding::Int: The dimension of the embedding.

max_length::Int: The maximum length of the sequence. Defaults to 1000.

Returns

PositionEncoding: A PositionEncoding object.

VDJDL.Embeddings.PositionEncoding — Type

PositionEncoding{W <: AbstractArray}(weight::W)

Represents positional encodings for sequence data.

Arguments

weight::W: The weight matrix for positional encodings.

VDJDL.Embeddings.PositionEncoding — Method

(pe::PositionEncoding)(x::AbstractArray)

Applies positional encoding to the input array.

Arguments

x::AbstractArray: The input array.

Returns

The positional encoding for the input array.

VDJDL.Embeddings.PositionEncoding — Method

(pe::PositionEncoding)(seq_length::Int)

Retrieves the positional encoding for a specific sequence length.

Arguments

seq_length::Int: The length of the sequence.

Returns

A view of the positional encoding matrix for the given sequence length.