VDJDL.Tokenizer.read_fasta — Method

read_fasta(filepath::String)

Reads a FASTA file and returns a vector of tuples containing the description and sequence of each record.

Arguments

filepath::String: The path to the FASTA file.

Returns

Vector{Tuple{String, String}}: A vector of tuples with the description and sequence.

VDJDL.Tokenizer.LabelEncoder — Type

LabelEncoder(labels::Vector{String})

Encodes and decodes labels.

Arguments

labels::Vector{String}: A vector of labels to be encoded.

Returns

LabelEncoder: A LabelEncoder object.

VDJDL.Tokenizer.SequenceTokenizer — Type

SequenceTokenizer(alphabet::Vector{T}, unksym::T) where T

Tokenizes character sequences.

Arguments

alphabet::Vector{T}: A vector of symbols to be tokenized.

unksym::T: The symbol for unknown tokens.

Returns

SequenceTokenizer{T}: A SequenceTokenizer object.