Base.rand — Method

Distributions.rand(rng::AbstractRNG, d::BiZISNB, n::Int)

Generates n random samples from the BiZISNB distribution.

Base.rand — Method

Distributions.rand(rng::AbstractRNG, d::BiZISNB)

Generates a random sample from the BiZISNB distribution.

Base.rand — Method

Distributions.rand(rng::AbstractRNG, d::MixtureZISNB)

Generates a random sample from the MixtureZISNB distribution.

Base.rand — Method

Distributions.rand(rng::AbstractRNG, d::ZISNB)

Generates a random sample from the Zero-Inflated Shifted Negative Binomial distribution.

Base.rand — Method

Base.rand(d::ShiftedPoisson)

Generates a random sample from the Shifted Poisson distribution.

Distributions.logpdf — Method

Distributions.logpdf(d::BiZISNB, x::Tuple{Int, Int})

Calculates the log of the probability density function for the BiZISNB distribution at a given tuple x.

Distributions.logpdf — Method

Distributions.logpdf(d::MixtureZISNB, x::Real)

Calculates the log of the probability density function for the MixtureZISNB distribution at a given value x.

Distributions.logpdf — Method

Distributions.logpdf(d::ShiftedPoisson, x::Int)

Calculates the log of the probability density function for the Shifted Poisson distribution at a given integer x.

Distributions.logpdf — Method

Distributions.logpdf(d::ZISNB, x::Int)

Calculates the log of the probability density function for the Zero-Inflated Shifted Negative Binomial distribution at a given integer x.

Distributions.pdf — Method

Distributions.pdf(d::ShiftedPoisson, x::Int)

Calculates the probability density function for the Shifted Poisson distribution at a given integer x.

StatsAPI.loglikelihood — Method

Distributions.loglikelihood(d::BiZISNB, x::Tuple{Int, Int})

Calculates the log likelihood for the BiZISNB distribution at a given tuple x.

VDJDL.VDJDistributions.BiZISNB — Type

BiZISNB{T1, T2, L}(π_left, p_left, π_right, p_right, α, β, len)

Defines the BiZISNB distribution that combines two ZISNB distributions.

π_left: Probability of zero inflation for the left trimming.

p_left: Probability of success for the left trimming.

π_right: Probability of zero inflation for the right trimming.

p_right: Probability of success for the right trimming.

α: Coefficient for the exponential part of the distribution.

β: Coefficient for the length-dependent part of the distribution.

len: Length parameter.

VDJDL.VDJDistributions.MixtureZISNB — Type

MixtureZISNB{T}(zisnb, negative_min, negative_max, mixture_prob)

Defines the MixtureZISNB distribution that combines the ZISNB and DiscreteUniform distributions for negative values.

zisnb: The Zero-Inflated Shifted Negative Binomial component.

negative_min: Minimum value for the negative range.

negative_max: Maximum value for the negative range.

mixture_prob: Probability of selecting the ZISNB component.

VDJDL.VDJDistributions.ShiftedPoisson — Type

ShiftedPoisson{T}(λ, shift)

Defines the Shifted Poisson distribution.

λ: The rate parameter of the Poisson distribution.

shift: The integer value by which the Poisson distribution is shifted.

VDJDL.VDJDistributions.ZISNB — Type

ZISNB{T}(π, r, p)

Defines the Zero-Inflated Shifted Negative Binomial distribution.

π: Probability of zero inflation.

r: Number of successes for the negative binomial component.

p: Probability of success for the negative binomial component.