Base.show — Method

Show the Rezero layer in a human-readable format.

Arguments

io::IO: The I/O stream.

m::Rezero: The Rezero layer instance.

Base.show — Method

Base.show(io::IO, r::RotaryMultiHeadAttention)

Custom string representation for RotaryMultiHeadAttention.

Arguments

io::IO: The IO stream.

r::RotaryMultiHeadAttention: An instance of RotaryMultiHeadAttention.

Example

r = RotaryMultiHeadAttention(256, 64, 8)
println(r)

VDJDL.Layers.Rezero — Type

Rezero layer implementation for Flux.jl. This layer multiplies the input by a learnable parameter alpha. Note this layer leads to higher loss value as comapred to LayerNorm given everything else constant in our experiments, but it is faster.

Fields

alpha::AbstractVector: The learnable parameter initialized by the user.

Constructor

Rezero(init::AbstractVector): Create a new Rezero layer with the specified initial values for alpha.

Rezero(): Create a new Rezero layer with alpha initialized to zeros of type Float32 with length 1.

Methods

(a::Rezero)(x::AbstractArray): Apply the Rezero layer to the input array x.

VDJDL.Layers.Rezero — Method

Apply the Rezero layer to the input array x.

Arguments

x::AbstractArray: The input array.

Returns

AbstractArray: The transformed array after applying the Rezero layer.

VDJDL.Layers.Rezero — Method

Create a new Rezero layer with alpha initialized to zeros of type Float32 with length 1.

Returns

Rezero: A new instance of the Rezero layer.

VDJDL.Layers.RotaryMultiHeadAttention — Type

struct RotaryMultiHeadAttention

A struct representing a multi-head attention mechanism with rotary position embeddings.

Fields

mha::MultiHeadAttention: An instance of MultiHeadAttention from Flux.

VDJDL.Layers.RotaryMultiHeadAttention — Method

(r::RotaryMultiHeadAttention)(q_in, k_in, v_in; bias=nothing, mask=nothing) -> Tuple

Apply the rotary multi-head attention mechanism to the input queries, keys, and values.

Arguments

r::RotaryMultiHeadAttention: An instance of RotaryMultiHeadAttention.

q_in: Input queries.

k_in: Input keys.

v_in: Input values.

bias: Optional bias for the attention mechanism.

mask: Optional mask for the attention mechanism.

Returns

A tuple (x, α) where x is the output of the attention mechanism and α is the attention weights.

VDJDL.Layers.RotaryMultiHeadAttention — Method

RotaryMultiHeadAttention(hidden_size::Int, head_size::Int, nheads::Int; bias::Bool=false, dropout_prob::Float64=0.0) -> RotaryMultiHeadAttention

Create a RotaryMultiHeadAttention instance.

Arguments

hidden_size::Int: The size of the hidden layer.

head_size::Int: The size of each attention head.

nheads::Int: The number of attention heads.

bias::Bool: Whether to use bias in the attention layers (default: false).

dropout_prob::Float64: Dropout probability for the attention mechanism (default: 0.0).

Returns

A RotaryMultiHeadAttention instance.